fix: keep engines handler outside route module
